feat: user login A user when logs in is given two types of token. An Access token A Refresh Token.
The Access token is used for 30 minutes and then expires.
The refresh token is then used continuously. This is important
becasue we do not want our access token to stay that long and also
when a user wants to do something critical like update password,
they request for a new access token.